#56 merged all examples in one console project that allow selection
