+++ aiv-fast2d-example/Alien/Assets/2.ico
