+++ aiv-fast2d-example/Alien/Assets/owl.png
