+++ aiv-fast2d-example/Alien/Scripts/MaskEffect.cs
