+++ aiv-fast2d-example/aiv-fast2d-example.csproj
@@ -42,22 +42,26 @@
     <Reference Include="Microsoft.CSharp" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Tilemap.cs" />
-    <Compile Include="WobbleEffect.cs" />
-    <Compile Include="WASDEffect.cs" />
-    <Compile Include="MaskEffect.cs" />
-    <Compile Include="RedBands.cs" />
-    <Compile Include="BlackBands.cs" />
-    <Compile Include="GrayscaleEffect.cs" />
-    <Compile Include="ParticleSystem.cs" />
+    <Compile Include="Alien\Scripts\Tilemap.cs" />
+    <Compile Include="Alien\Scripts\WobbleEffect.cs" />
+    <Compile Include="Alien\Scripts\WASDEffect.cs" />
+    <Compile Include="Alien\Scripts\MaskEffect.cs" />
+    <Compile Include="Alien\Scripts\RedBands.cs" />
+    <Compile Include="Alien\Scripts\BlackBands.cs" />
+    <Compile Include="Alien\Scripts\GrayscaleEffect.cs" />
+    <Compile Include="Alien\Scripts\ParticleSystem.cs" />
+    <Compile Include="Alien\AlienExample.cs" />
+    <Compile Include="DrawTexture\DrawTextureExample.cs" />
+    <Compile Include="MultiWindow\MultiWindowExample.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Rope.cs" />
-    <Compile Include="Segment.cs" />
+    <Compile Include="Alien\Scripts\Rope.cs" />
+    <Compile Include="Alien\Scripts\Segment.cs" />
+    <Compile Include="RenderTexture\RenderTextureExample.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
-    <None Include="Assets\map001.csv">
+    <None Include="Alien\Assets\map001.csv">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
     <None Include="OpenTK.dll.config" />
@@ -70,28 +74,38 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Assets\LogoAIV.png" />
+    <EmbeddedResource Include="Alien\Assets\LogoAIV.png" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Assets\2.png" />
+    <EmbeddedResource Include="Alien\Assets\2.png" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Assets\owl.png" />
+    <EmbeddedResource Include="Alien\Assets\owl.png" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Assets\2.ico" />
+    <EmbeddedResource Include="Alien\Assets\2.ico" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Assets\mask_circle.png" />
+    <EmbeddedResource Include="Alien\Assets\mask_circle.png" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Assets\mask_circle2.png" />
+    <EmbeddedResource Include="Alien\Assets\mask_circle2.png" />
   </ItemGroup>
   <ItemGroup>
-    <Content Include="Assets\tiles_spritesheet.png">
+    <Content Include="Alien\Assets\tiles_spritesheet.png">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Content>
+    <Content Include="DrawTexture\Assets\aj.jpg">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <None Include="MultiWindow\Assets\circle.png">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <Content Include="RenderTexture\Assets\star.png">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
