+++ aiv-fast2d.sln
@@ -7,12 +7,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "aiv-fast2d", "aiv-fast2d\ai
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "aiv-fast2d-example", "aiv-fast2d-example\aiv-fast2d-example.csproj", "{051FC68A-81B7-46AA-BADD-BFDEA23E9476}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "aiv-fast2d-example-units", "aiv-fast2d-example-units\aiv-fast2d-example-units.csproj", "{5A87C0B2-EDC9-45A5-AA5F-0FA1502FE79B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "aiv-fast2d-getpixels", "aiv-fast2d-getpixels\aiv-fast2d-getpixels.csproj", "{387D495A-FF22-4D6B-A037-3F7D58CEA46F}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "aiv-fast2d-check-texture", "aiv-fast2d-check-texture\aiv-fast2d-check-texture.csproj", "{DF979CC3-FBD9-4889-8BD2-0EDFF76CEF69}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,18 +21,6 @@ Global
 		{051FC68A-81B7-46AA-BADD-BFDEA23E9476}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{051FC68A-81B7-46AA-BADD-BFDEA23E9476}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{051FC68A-81B7-46AA-BADD-BFDEA23E9476}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5A87C0B2-EDC9-45A5-AA5F-0FA1502FE79B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5A87C0B2-EDC9-45A5-AA5F-0FA1502FE79B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5A87C0B2-EDC9-45A5-AA5F-0FA1502FE79B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5A87C0B2-EDC9-45A5-AA5F-0FA1502FE79B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{387D495A-FF22-4D6B-A037-3F7D58CEA46F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{387D495A-FF22-4D6B-A037-3F7D58CEA46F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{387D495A-FF22-4D6B-A037-3F7D58CEA46F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{387D495A-FF22-4D6B-A037-3F7D58CEA46F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DF979CC3-FBD9-4889-8BD2-0EDFF76CEF69}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DF979CC3-FBD9-4889-8BD2-0EDFF76CEF69}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DF979CC3-FBD9-4889-8BD2-0EDFF76CEF69}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DF979CC3-FBD9-4889-8BD2-0EDFF76CEF69}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
